requirements.txt
h5py==3.2.1
joblib==1.0.1
Keras==2.4.3
numpy==1.20.3
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
threadpoolctl==2.1.0

pandas~=1.2.4
nltk~=3.6.2
tweety~=0.1.6
html5lib~=1.1
requests~=2.25.1
bs4~=0.0.1
beautifulsoup4~=4.9.3